+++ app-latest/src/pages/tabs/content/newInc/newIncStepByStep/step1/step1.ts
@@ -1,6 +1,7 @@
 import {Component, forwardRef, NgZone, Provider} from '@angular/core';
 import {NavController, NavParams, MenuController, ViewController, 
         Platform, Events} from 'ionic-angular';
+import { FormControl } from '@angular/forms';
 //import {AndroidAttribute} from './../../../../../../directives/global.helpers';
 import {ConferenceData} from './../../../../../../providers/conference-data';
 import {marker} from './../../newIncInterface';
@@ -17,6 +18,7 @@ import {defaultLanguage, folderLanguage, sourceLanguage, compareLanguage, useSQL
 import {Step1SearchPipe} from './step1Pipe';
 import {Step2Page} from './../step2/step2';
 import {Storage} from '@ionic/storage';
+import 'rxjs/add/operator/debounceTime';
 
 @Component({
   selector: 'ste1-page',
@@ -30,8 +32,11 @@ export class Step1Page {
   user: any = {};
   //storage: any;
   searchText: any;
+  searchControl: FormControl;
   inc: any;
   tiposElementos: any;
+  tiposElementosOriginal: any;
+  searching: any = false;
   
   constructor(private platform: Platform
     , private menu: MenuController
@@ -50,6 +55,8 @@ export class Step1Page {
       this.inc = params.data;
       this.searchText = '';
 
+      this.searchControl = new FormControl();
+
       //this.storage = new Storage(SqlStorage);  
 
       this.storage.get('user').then((user) => {
@@ -63,14 +70,35 @@ export class Step1Page {
                 this.tiposElementos = this.tiposElementos.filter(item => item.FamiliaTipoElementoID == this.inc.familia.FamiliasTiposElementosID
                 && item.EsInterno == false);
               }
-              
+              this.tiposElementosOriginal = this.tiposElementos;
           });
       });
     
   }
 
-  inputSearch(search) {
+  ionViewDidLoad() {
+ 
+        this.searchControl.valueChanges.debounceTime(700).subscribe(search => { 
+            this.searching = false;
+            this.tiposElementos = this.filter(this.tiposElementosOriginal, search); 
+        });
+ 
+  }
+
+  filter(value, args){
+    if (args){
+      if (args.toLowerCase() != '') {
+        return value.filter((item)=>
+            item.DesTipoElemento.toLowerCase().indexOf(args.toLowerCase()) != -1
+        );
+      }
+    }
+    return value;
+  }
+
+  inputSearch($event) {
     //console.log(search.value);
+    this.searching = true;
   }
 
   openIncidents(element){